[hotfix/#21] 회원 등록 후 홈에서의 식물 GET이 바로 이루어지지 않는 문제 해결
